plugins/yolo/pyproject.toml
[tool.poetry]
name = "tableai-yolo-service"
version = "0.1.0"
description = "Standalone YOLO table detection service"
authors = ["Schlafdata <john@schlafdata.com>"]
readme = "README.md"
license = ""

packages = [
    {include = "app"},
    {include = "tableai_tools", from = "tools"}
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"

# Core tools dependencies (will be included from tools/tableai_tools)
pandas = "2.2.3"
numpy = "2.2.4"
PyMuPDF = "1.25.5" 
requests = "2.32.3"
json-repair = "0.32.0"
pydantic = "2.9.2"
pyyaml = "6.0.1"
boto3 = "1.38.24"
pytesseract = "0.3.13"
ocrmypdf = "16.10.1"
pdf2image = "1.17.0"
Pillow = "^10.0.0"
ultralyticsplus = "0.1.0"
ultralytics = "8.0.225"
ipython = "8.37.0"
dateparser = "1.2.1"

# FastAPI service dependencies
fastapi = "^0.115.4"
uvicorn = {extras = ["standard"], version = "^0.29"}
python-multipart = "^0.0.6"

# HuggingFace for model downloads
huggingface-hub = "^0.20.0"

[tool.poetry.extras]
cpu = ["torch", "torchvision", "ultralytics"]
gpu = ["torch", "torchvision", "ultralytics"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"